rename file to match hook name
